started on the test suite after seeding location data and setting up db connection.
